Extract test circuit mock to a separate file
